Remove Pepsi Corporate project and rename files
