+++ src/constants/type-inference.ts
